Fix issue with FB UIDs
Thought functionality integers, pyFacebook returns the UID as a string type.
The string replacement for creating a new user assumed it was getting an
integer and Python won't convert the string without explicitly doing so.

This modified code tests the string case in addition to the interger and makes
sure they both get through.